Isolate phone-related capabilities to a plugins subdirectory
