+++ docs/installation.rst
@@ -34,6 +34,7 @@ Add the following apps to the ``INSTALLED_APPS``:
         'django_otp.plugins.otp_static',
         'django_otp.plugins.otp_totp',
         'two_factor',
+        'two_factor.plugins.phonenumber',  # <- if you want phone number capability.
     )
 
 Add the ``django-otp`` middleware to your ``MIDDLEWARE``. Make sure
