+++ example/settings.py
@@ -63,6 +63,7 @@ INSTALLED_APPS = (
     'django_otp.plugins.otp_static',
     'django_otp.plugins.otp_totp',
     'two_factor',
+    'two_factor.plugins.phonenumber',
     'example',
 
     'debug_toolbar',
