+++ two_factor/templatetags/two_factor.py
@@ -4,8 +4,6 @@ import phonenumbers
 from django import template
 from django.utils.translation import gettext as _
 
-from ..models import PhoneDevice
-
 register = template.Library()
 
 phone_mask = re.compile('(?<=.{3})[0-9](?=.{2})')
@@ -43,14 +41,14 @@ def format_phone_number(number):
 @register.filter
 def device_action(device):
     """
-    Generates an actionable text for a :class:`~two_factor.models.PhoneDevice`.
+    Generates an actionable text for a :class:`~two_factor.plugins.phonenumber.models.PhoneDevice`.
 
     Examples:
 
     * Send text message to `+31 * ******58`
     * Call number `+31 * ******58`
     """
-    assert isinstance(device, PhoneDevice)
+    assert device.__class__.__name__ == 'PhoneDevice'
     number = mask_phone_number(format_phone_number(device.number))
     if device.method == 'sms':
         return _('Send text message to %s') % number
