+++ two_factor/views/__init__.py
@@ -1,6 +1,5 @@
 from .core import (
-    BackupTokensView, LoginView, PhoneDeleteView, PhoneSetupView,
-    QRGeneratorView, SetupCompleteView, SetupView,
+    BackupTokensView, LoginView, QRGeneratorView, SetupCompleteView, SetupView,
 )
 from .mixins import OTPRequiredMixin
 from .profile import DisableView, ProfileView
